Moved line break validation to separate file
